refactor : ZWRC-47 : 모델 파일 분리 작업 완료
