+++ api/workflow_engine_app/models/workflow.py
@@ -1,0 +1,10 @@
+from django.db import models
+import uuid
+
+
+class Workflow(models.Model):
+    uuid = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
+    name = models.CharField(max_length=255)
+    description = models.TextField()
+    created_at = models.DateTimeField(auto_now_add=True)
+    updated_at = models.DateTimeField(auto_now=True)
